Perform global deopt lazily as we unwind callstack
This is a bit more CPU cache friendly as it means we don't need to go
looking up a bunch of stuff relating to a frame down the stack, only the
record itself. The primary reason for this, however, is that it means we
shall only have to perform uninlining at the very top of the call stack,
which greatly eases the process (since it's trivial to create records at
the top, but would be very fraught, and quite costly, to rewrite the
whole stack at the point of a global deopt). The update of uninlining is
still to come.