Nothing reads $config->{can_err_decl_after_stmt} so don't set it.
This line was added in Sept 2017 as part of commit 23c16b303163c621:
    Probe for gcc -Werror=* support. This allows building MoarVM on older GCCs.

    The -Werror=* probe has to run before setting the cc/ldmiscflags. The
    compiler_usability check only makes sense before the -Werror=* probe. Thus
    that one is also moved earlier.

However that commit didn't add code which read it, and as best I can tell
nothing else ever did either.

The "sense" of the name even looks backwards - if the compiler *isn't* gcc,
then surely it *can't* accept `-Werror=declaration-after-statement`?