Add a probe for log() returning NaN for negative input.
Similar to commit 5e031c917587f8b3 for pow(), Solaris also seems to show
non-POSIX behaviour for log() and log10(), for the same compilers.

Having written the tests for both log() and log10(), I realise that we
don't actually use log10(). As the work is done, doesn't add much
complexity, and we might use log10() in the future, leave the code in place.