Move definitions of dll{import,export,local} from 'Win32' to the toolchains.
Every other definition of these three setup keys is in a "toolchain", not
an "OS".

This requires us to create a 'cygwin' toolchain, to override these 3 keys
from the 'gcc' defaults to keep the generated configuration identical.

I cannot verify that 'cygwin' (or 'mingw32' for that matter) builds
correctly. I have tested that this change (and none on this branch) do not
cause any changes to any of the generated files for the default
invocation of ./Configure.pl --os=$os

ie if it worked before, it will still work. If it was broken before, it will
be broken the same way as before. But it is now easier to fix problems with
other parts of the build system.

I am suspicious that a lot more configuration keys should be the same for
'cygwin' and 'mingw32'. Commit 8c2c627ca5fe1a13 in Feb 2017 reports that
the cygwin build was broken - it's unclear when it last worked, as no other
commits change any part of it since it was first added in 2013. MinGW
configuration was updated in Feb 2017, implying that it was confirmed
working at that time.