Allow outer init args to be used for inner ones
Consider a situation where we have a method dispatcher and a wrap
dispatcher, all doing `callsame`. The args to the initial dispatch are
saved as the resume initialization state, so that we'll have access to
them at resume time. The wrap dispatcher works through the wrappers, and
eventually is exhaused. It thus falls back to the method dispatcher,
which in turn reaches for the next method and dispatches to it, which
starts a new wrap dispatch. That wrap dispatch is also resumable so
wants to set resumption initialization args. Those were themselves
obtained from resumption initialization args - a situation that was not
previously permitted.

To support it, we copy those args into temporaries. This is the safest
way, and means we don't have to do a delegate dance of trying to work
out where the arguments were originally have sourced from, which would
be fragile. The bookkeeping cost would likely also be higher than the
copying cost anyway. The main thing we want to achieve by keeping the
resumption initialization args to zero cost per dispatch is for the very
common case where we didn't resume; if we are already in the midst of a
relatively complex resumption, it's fair game that there's a small cost
there.